Reduce min time between replies to 20 seconds
Idea is to find the time that replies to all human sent messages but only
avoid infinite loops. 20 second delay should probably be fine. 20 second
delay is also a little too slow for instant messaging. Well it is possible
that in some cases it may take longer, but I am confident it breaks the loop
at least at 5th or 10th message.